chore(banks): move bank resources into banks subfolder
