+++ app/Resource/Banks/Privatbank/PrivatbankCurrencyExchangeRateResource.php
@@ -2,16 +2,13 @@
 
 declare(strict_types=1);
 
-namespace App\Resource\Monobank;
+namespace App\Resource\Banks\Privatbank;
 
-use Carbon\CarbonImmutable;
-
-final readonly class MonobankCurrencyExchangeRateResource
+final readonly class PrivatbankCurrencyExchangeRateResource
 {
     public function __construct(
         public string $currencyA,
         public string $currencyB,
-        public CarbonImmutable $dateTime,
         public float $rateSell,
         public float $rateBuy,
     ) {
